feat(fomo-web): 스플래시 제거(바로 메인) + 로그인 바텀시트→페이지 + 카드클릭 로그인 게이트
광혁 지시 2건:
1) 스플래시("나만 그런 거 아냐") 제거 — 진입 즉시 메인. 데이터는 백그라운드 로드,
   도착 전엔 HomeView 내부 로딩 상태가 담당. page.tsx phase 에서 splash 제거(home 시작).
2) 로그인 = 바텀시트(AuthSheet) → 풀페이지(LoginPage). 메인에서 카드 클릭 시(비로그인)
   뎁스 대신 로그인 페이지로(KeywordCardFeed openItem/openInterest 게이트). 로그인 상태면 기존대로 뎁스.

AuthSheet 제거(LoginPage 로 대체). SplashScreen 컴포넌트 파일은 보존(미사용).
검증: typecheck·build green. 프리뷰 — 진입 즉시 메인, 카드 탭→풀페이지 로그인 확인.
